[upd] Less-overflowable operator*= to Fp<m>
Previous version of Fp<m> could overflow and return a wrong result when
the product of the two numbers exceeds the limit of Fp<m>::number_type.

This version detects overflowing cases and switches the multiplication
algorithm to a less-overflowable one. The less-overflowable multiplication
has the time-complexity of O(logN) where N is the right-operand of
multiplication, and can handle all cases as long as (the left-operand)
+= (the left-operand itself) does not overflow.